🔨 refactor: 마우스 호버 이벤트 훅 관리 + 라인클래스이름 추상화
