+++ apps/client/src/shared/code-highlighter/components/LineNumbers/LineNumbers.tsx
@@ -1,5 +1,6 @@
 import styles from '../../styles/CodeViewer.module.css';
-import { useState } from 'react';
+import { useHoverLineNumber } from '../../hooks/useHoverLineNumber';
+import { getLineNumberClass } from '../../utils/getLineNumberClass';
 
 type LineNumbersProps = {
   codeLineList: string[];
@@ -11,16 +12,7 @@ type LineNumbersProps = {
  * 코드의 줄 수를 표시하는 컴포넌트
  */
 export const LineNumbers = ({ codeLineList }: LineNumbersProps) => {
-  const [hoveredLineNumber, setHoveredLineNumber] = useState<number | null>(null);
-
-  // 마우스 enter, leave 색상 변화
-  const handleMouseEnter = (lineNumber: number) => {
-    setHoveredLineNumber(lineNumber);
-  };
-
-  const handleMouseLeave = () => {
-    setHoveredLineNumber(null);
-  };
+  const { hoveredLineNumber, handleMouseEnter, handleMouseLeave } = useHoverLineNumber();
 
   return (
     <div className={styles.lineNumbers}>
@@ -29,7 +21,12 @@ export const LineNumbers = ({ codeLineList }: LineNumbersProps) => {
           key={index}
           onMouseEnter={() => handleMouseEnter(index + 1)}
           onMouseLeave={handleMouseLeave}
-          className={`${styles.lineNumber} ${hoveredLineNumber === index + 1 ? styles.lineHighlight : ''}`}
+          className={getLineNumberClass(
+            styles.lineNumber,
+            styles.lineHighlight,
+            index + 1,
+            hoveredLineNumber
+          )}
         >
           {index + 1}
         </div>
