+++ apps/client/src/shared/code-highlighter/utils/getLineNumberClass.ts
@@ -1,0 +1,8 @@
+export const getLineNumberClass = (
+  baseClass: string,
+  highlightClass: string,
+  lineNumber: number,
+  hoveredLineNumber: number | null
+): string => {
+  return `${baseClass} ${hoveredLineNumber === lineNumber ? highlightClass : ''}`.trim();
+};
